+++ git-repository/tests/easy/ext/mod.rs
@@ -1,0 +1,2 @@
+mod object;
+mod reference;
