+++ git-repository/tests/easy/mod.rs
@@ -1,3 +1,4 @@
 mod access;
+mod ext;
 mod object;
 mod reference;
